Fixed issue with object paging so see all objects in bucket (the issue is still probably in the bucket paging).
Added a sort to the treeview.

Added a filter input text field that can be used to filter the objects in the buckets when the refresh buckets button is pressed.

Added a show manifest button so you can see the object status, and more importantly, the reason if the translate failed.